Camera movemement

+++ code/headers/Color_Buffer_Rgb565.hpp
@@ -76,7 +76,7 @@
 
             int size () const
             {
-                return (buffer.size ());
+                return ((int)buffer.size ());
             }
 
         public:
