+++ code/headers/Color_Buffer_Rgba8888.hpp
@@ -91,7 +91,7 @@
 
             int size () const
             {
-                return (buffer.size ());
+                return ((int)buffer.size ());
             }
 
         public:
